Rework model loading
Instead of loading directly the functions from a modules as entry
points (as we were doing) we can just point the entrypoint to a module
or object that implements some methods, namely:

    predict_data
    predict_file
    predict_url
    metadata

Then we use an object to handle the calls to those functions, catching
exceptions and returning 405 errors if they are not implemented.

This is still WIP, as no documentation is done. Internal APIs are not
defined yet.