Allow uploading files as well as pointing to URLs
We may allow to pass several arguments at once in order to make better
predictions. We thus allow passing several URLs to load the files for
the inference. However, currently (and unfortunately) we allow only to
upload one file. There is a bug in the Swagger UI that makes impossible
to upload several files, although this works if the request is not done
through swagger. Since providing a UI and an API that behave differently
is incoherent, only allow one file for the time being.

See https://github.com/noirbizarre/flask-restplus/issues/491 for more
details on the bug.