Initial support for passing down training arguments
Currently we were not able to pass down training arguments to the model.
We could simply pass down all the arguments that we receive, but the
problem is that we cannot do simply like this, as it would be a
nightmare for users to figure out which arguments are expected.

This change implements the loading of arguments for each model. Since
different models would have different arguments, we need to pre-populate
the API with those routes.