Fix test arguments
This commit is a patch to be consistent with the previous version of
DEEPaaS, in which we returned a list of len 1 (because of the Swagger
bug) for the args['files'] parameter in the predict_data method.

I forgot to add this in the test-args branch and was returning directly
a FileStorage object (not a list of FileStorage objects).