Use asyncio in ModelWrapper and adapt API
After moving to aiohttp it is easy to move the calls to de underlying
method to coroutines using asyncio. In order to do so we convert the
ModelWrapper to use async methods, and then we adapt the API to these
changes. As a side effect, we change the way we are defining the routes
and views, since we need to include a configuration option defining the
number of worker instances that we want to spawn.

We also implement a new warm method that can be defined by the user, so
that this method is called before the API is spawned.

Sem-Ver: api-break