Set basePath on swagger.json when serving under OpenWhisk
We need to setup the proper base path when we are running under
OpenWhisk, as the OpenAPI spec states that all URLs are constructed
using basePath (if exists) or otherwise they are absolute urls. Since we
are running as Web actions in OW, we need to set the basePath to the
actual action path.