Pin webargs version <6.0.0
Webargs >= 6.0.0 introduced a breaking change where schema fields may
not specify a location any longer. We need to change our code to specify
per-location schemas (easy) but aiohttp-apispec is not ready yet until
maximdanilchenko/aiohttp-apispec/83 is fixed.

This change pins the webargs version to something below 6.0.0 so that
things are not borken.

Fixes #82
Sem-Ver: bugfix