Unify train and predict pools to fix GPU OOM
This fixes GPU out-of-memory problems that happened when we had two
different pools (for predict and train). When we did train then
predict sequentially (or viceversa) each pool wanted to have the whole
GPU so out-of-memory errors happened. This won't fix out-of-memory
errors when running parallel tasks on GPU (errors which also happened
before).

CPU deployments shouldn't be affected.

Fixes #87